Taha mirza

+++ src/app/api/mentor/proposal/delete/route.js
@@ -11,7 +11,7 @@ export async function DELETE(request) {
         const { proposalID } = await request.json();
         const profileID = request.headers.get('profileID');
         const proposal = await Proposal.findById(proposalID);
-        if (proposal.edit) {
+        if (proposal.edit && proposal.proposedBy == 'Mentor' && proposal.proposer == profileID) {
             const deletedProposal = await Proposal.findByIdAndDelete(proposalID);
             const mentor = await Mentor.findById(profileID);
             const myProposals = mentor.myProposals.filter(proposal => proposal != proposalID);
